ani kokot nejde

+++ mi6maven/src/main/java/managers/AssignmentManagerImpl.java
@@ -8,14 +8,20 @@ package managers;
 import entities.Agent;
 import entities.Assignment;
 import entities.Mission;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.sql.DataSource;
 import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.PreparedStatementCreator;
 import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
 
 /**
  *
@@ -30,9 +36,11 @@ public class AssignmentManagerImpl implements AssignmentManager {
     private static MissionManagerImpl missionManager = new MissionManagerImpl();
 
     private static final Logger logger = Logger.getLogger(AssignmentManagerImpl.class.getName());
-
+    
     public void setDataSource(DataSource dataSource) {
         this.jdbc = new JdbcTemplate(dataSource);
+        agentManager.setDataSource(dataSource);
+        missionManager.setDataSource(dataSource);
     }
 
     private static final RowMapper<Assignment> MAPPER = new RowMapper<Assignment>() {
@@ -49,21 +57,58 @@ public class AssignmentManagerImpl implements AssignmentManager {
     };
 
     @Override
-    public void createAssignment(Assignment assignment) {
+    public void createAssignment(final Assignment assignment) {
         validateAssignment(assignment);
-        jdbc.update("INSERT INTO ASSIGNMENTS (MISSION_ID, AGENT_ID, START_DATE, END_DATE) VALUES (?,?,?,?)",
-                assignment.getAgent().getId(), assignment.getMission().getId(), assignment.getStartDate(), assignment.getEndDate());
+
+        if (assignment.getEndDate() != null && assignment.getEndDate().before(assignment.getStartDate())) {
+            String message = "wrong dates";
+            logger.log(Level.WARNING, message);
+            throw new IllegalArgumentException(message);
+        }
+
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+
+        jdbc.update(
+                new PreparedStatementCreator() {
+                    @Override
+                    public PreparedStatement createPreparedStatement(Connection connection) throws SQLException {
+                        PreparedStatement ps
+                        = connection.prepareStatement("INSERT INTO ASSIGNMENTS (MISSION_ID, AGENT_ID, START_DATE, END_DATE) VALUES (?,?,?,?)", Statement.RETURN_GENERATED_KEYS);
+                        ps.setLong(1, assignment.getAgent().getId());
+                        ps.setLong(2, assignment.getMission().getId());
+                        ps.setDate(3, new java.sql.Date(assignment.getStartDate().getTime()));
+                        ps.setDate(4, new java.sql.Date(assignment.getEndDate().getTime()));
+                        return ps;
+                    }
+                },
+                keyHolder);
+        assignment.setId(keyHolder.getKey().longValue());
+
     }
 
     @Override
     public void updateAssignment(Assignment assignment) {
         validateAssignment(assignment);
+        Assignment assignmentFromDB = jdbc.queryForObject("SELECT * FROM ASSIGNMENTS WHERE ID=?", MAPPER, assignment.getId());
+
+        if (assignmentFromDB.getEndDate() != null && !assignmentFromDB.getEndDate().equals(assignment.getEndDate())) {
+            String message = "change of set end date is forbidden";
+            logger.log(Level.WARNING, message);
+            throw new IllegalArgumentException(message);
+        }
+
         jdbc.update("UPDATE ASSIGNMENTS SET AGENT_ID=? MISSION_ID=?, START_DATE=?, END_DATE=?  WHERE ID=?",
                 assignment.getAgent().getId(), assignment.getMission().getId(), assignment.getStartDate(), assignment.getEndDate(), assignment.getId());
     }
 
     @Override
     public void deleteAssignment(Assignment assignment) {
+        if (assignment == null) {
+            String message = "deleting null assignment";
+            logger.log(Level.WARNING, message);
+            throw new IllegalArgumentException(message);
+        }
+
         jdbc.update("DELETE FROM ASSIGNMENTS WHERE id=?", assignment.getId());
     }
 
@@ -79,12 +124,12 @@ public class AssignmentManagerImpl implements AssignmentManager {
 
     @Override
     public List<Assignment> findAssignmentsForAgent(Agent agent) {
-        return jdbc.query("SELECT * FROM ASSIGNMENTS WHERE AGENT_ID=?",MAPPER, agent.getId());
+        return jdbc.query("SELECT * FROM ASSIGNMENTS WHERE AGENT_ID=?", MAPPER, agent.getId());
     }
 
     @Override
     public List<Assignment> findAssignmentsForMission(Mission mission) {
-        return jdbc.query("SELECT * FROM ASSIGNMENTS WHERE MISSION_ID=?",MAPPER, mission.getId());
+        return jdbc.query("SELECT * FROM ASSIGNMENTS WHERE MISSION_ID=?", MAPPER, mission.getId());
     }
 
     private void validateAssignment(Assignment assignment) {
@@ -93,7 +138,7 @@ public class AssignmentManagerImpl implements AssignmentManager {
             logger.log(Level.WARNING, message);
             throw new IllegalArgumentException(message);
         }
-        
+
         if (assignment.getAgent() == null) {
             String message = "agent cannot be null";
             logger.log(Level.WARNING, message);
